Move ``typhon.eom`` -> ``typhon.retrieval.oem``
Make the OEM utility functions part of the retrieval subpackage.
