--- doc/typhon.oem.rst
@@ -1,16 +1,0 @@
-oem
-===
-
-.. automodule:: typhon.oem
-
-.. currentmodule:: typhon.oem
-
-.. autosummary::
-   :toctree: generated
-
-   error_covariance_matrix
-   averaging_kernel_matrix
-   retrieval_gain_matrix
-   smoothing_error
-   retrieval_noise
-
