+++ doc/typhon.retrieval.rst
@@ -31,6 +31,22 @@ retrieval.mcmc
 
    MCMC
 
+retrieval.oem
+=============
+
+.. automodule:: typhon.retrieval.oem
+
+.. currentmodule:: typhon.retrieval.oem
+
+.. autosummary::
+   :toctree: generated
+
+   error_covariance_matrix
+   averaging_kernel_matrix
+   retrieval_gain_matrix
+   smoothing_error
+   retrieval_noise
+
 retrieval.qrnn
 ==============
