+++ typhon/retrieval/oem/__init__.py
@@ -2,8 +2,8 @@
 """Collection of functions concerning the Optimal Estimation Method (OEM).
 """
 
-from typhon.oem.common import *  # noqa
-from typhon.oem.error import *  # noqa
+from typhon.retrieval.oem.common import *  # noqa
+from typhon.retrieval.oem.error import *  # noqa
 
 
 __all__ = [s for s in dir() if not s.startswith('_')]
